[Refactor] update smk variables, add smoke status cat5
Re-categorization of SMKDSTY to be used for smoking imputation. SMKDSTY_cat5 is a 5 category variable for smoking status for cycles up to 2018. Prior to 2015, 'occasional' and 'always occasional' are combined to form the current 'occasional' category. 2015 onwards, 'former occasional' and 'experimental' are combined to form the current 'former occasional' category.