fixed bug with ADL_01 variable
I believe this was a typo Doug had pointed out previously that he wanted me to
take a look into. The 2015-16 ADL_01 variable was reformatted to a 4 category
variable in these survey cycles so their inclusion in these rows would be
incorrect. Additional rows had been added to variable_details to account for
their harmonization back to a two category variable as well as rows to have
them coded as their original 4-category variable. There also should not be
2015-2016 or 2017-18 master cycles for these variables as the module wasn't
mandatory and therefore not collected in Ontario for those cycles.